+++ examples/corr_sdp.jl
@@ -36,11 +36,11 @@ m = Model(with_optimizer(CSDP.Optimizer))
 # Find upper bound
 @objective(m, Max, X[1,3])
 JuMP.optimize!(m)
-println("Maximum value is ", JuMP.result_value.(X)[1,3])
-@assert +0.8719 <= JuMP.result_value.(X)[1,3] <= +0.8720
+println("Maximum value is ", JuMP.value.(X)[1,3])
+@assert +0.8719 <= JuMP.value.(X)[1,3] <= +0.8720
 
 # Find lower bound
 @objective(m, Min, X[1,3])
 JuMP.optimize!(m)
-println("Minimum value is ", JuMP.result_value.(X)[1,3])
-@assert -0.9779 >= JuMP.result_value.(X)[1,3] >= -0.9799
+println("Minimum value is ", JuMP.value.(X)[1,3])
+@assert -0.9779 >= JuMP.value.(X)[1,3] >= -0.9799