+++ examples/qcp.jl
@@ -39,5 +39,5 @@ JuMP.optimize!(m)
 
 # Solution
 println("Objective value: ", JuMP.objective_value(m))
-println("x = ", JuMP.result_value(x))
-println("y = ", JuMP.result_value(y))
+println("x = ", JuMP.value(x))
+println("y = ", JuMP.value(y))